+++ example/dartactyl_example.dart
@@ -3,8 +3,10 @@ import 'package:dio/dio.dart';
 
 final dio = Dio();
 
-final apiKey = '<API-Key>';
-final panelUrl = 'https://panel.example.com';
+// final apiKey = '<API-Key>';
+// final panelUrl = 'https://panel.example.com';
+final panelUrl = 'https://panel.synahost.com';
+final apiKey = '8pPWzJhCkhhX3TYp5YYagvNhdmXRz3KJGjpEoUxcXGPwigGR';
 
 void main(List<String> args) async {
   print('Starting!');
@@ -15,8 +17,7 @@ void main(List<String> args) async {
   print('Getting Data!');
 
   await client.getServers().then((response) {
-    Server server = response.data[0].attributes;
-    print(server.name);
-    print(server.description);
+    Server? server = response.data?[0].attributes;
+    print('${server?.name} : ${server?.description}');
   });
 }