+++ lib/src/models/fractal/fractal_response_list.dart
@@ -1,13 +1,14 @@
-import '../../../models.dart';
-
 // json_serializable didnt work for me, so I did it myself
+import '../meta.dart';
+import 'fractal_response_data.dart';
+
 class FractalResponseList<T> {
   String object;
-  List<FractalResponseData<T>> data;
+  List<FractalResponseData<T>>? data;
   Meta? meta;
   FractalResponseList({
     required this.object,
-    required this.data,
+    this.data,
     this.meta,
   });
 
@@ -22,19 +23,27 @@ class FractalResponseList<T> {
 FractalResponseList<T> _$FractalResponseListFromJson<T>(
     Map<String, dynamic> json, T Function(Object? json) fromJsonT) {
   return FractalResponseList(
-    object: json['object'],
-    data: json['data']
-        .map<FractalResponseData<T>>(
-          (item) => FractalResponseData<T>(
-            object: item['object'],
-            attributes: fromJsonT(item['attributes'] as Map<String, dynamic>),
-            meta: item['meta'] == null
-                ? null
-                : Meta.fromJson(item['meta'] as Map<String, dynamic>),
-          ),
-        )
-        .toList(), // toList() is needed to convert the map to a list
-  );
+      object: json['object'],
+      data: () {
+        if (json['data'] != null && json['data'] is List) {
+          return (json['data'] as List)
+              .map<FractalResponseData<T>>(
+                (item) => FractalResponseData<T>(
+                  object: item['object'],
+                  attributes: item['attributes'] == null
+                      ? null
+                      : fromJsonT(item['attributes'] as Map<String, dynamic>),
+                  meta: item['meta'] == null
+                      ? null
+                      : Meta.fromJson(item['meta'] as Map<String, dynamic>),
+                ),
+              )
+              .toList();
+        } // toList() is needed to convert the map to a list
+        else {
+          return null;
+        }
+      }());
 }
 
 Map<String, dynamic> _$FractalResponseListToJson<T>(
@@ -44,7 +53,7 @@ Map<String, dynamic> _$FractalResponseListToJson<T>(
     <String, dynamic>{
       'object': instance.object,
       'data': instance.data
-          .map<Map<String, dynamic>>(
+          ?.map<Map<String, dynamic>>(
             (FractalResponseData<T> item) => item.toJson(toJsonT),
           )
           .toList(),