+++ lib/src/models/request_models/schedule.g.dart
@@ -7,8 +7,9 @@ part of 'schedule.dart';
 // **************************************************************************
 
 Schedule _$ScheduleFromJson(Map<String, dynamic> json) => Schedule(
+      onlyWhenOnline: json['only_when_online'] as bool,
       name: json['name'] as String,
-      isActive: json['is_active'] as bool?,
+      isActive: json['is_active'] as bool,
       minute: json['minute'] as String,
       hour: json['hour'] as String,
       dayOfWeek: json['day_of_week'] as String,
@@ -18,6 +19,7 @@ Schedule _$ScheduleFromJson(Map<String, dynamic> json) => Schedule(
 Map<String, dynamic> _$ScheduleToJson(Schedule instance) => <String, dynamic>{
       'name': instance.name,
       'is_active': instance.isActive,
+      'only_when_online': instance.onlyWhenOnline,
       'minute': instance.minute,
       'hour': instance.hour,
       'day_of_week': instance.dayOfWeek,